refactor: extract compile context template helpers and register e2e shard ownership
