+++ Makefile
@@ -57,6 +57,7 @@ E2E_DUCKDB_PYTEST_ARGS := -m "$(E2E_DUCKDB_MARKERS)" -vv --color=yes -n auto --d
 
 E2E_DUCKDB_BUILD_CORE_PATHS := \
 	tests/e2e/src/sqlbuild/cli/commands/main/build/no_tests_no_audits \
+	tests/e2e/src/sqlbuild/cli/commands/main/build/test_audit_factory_failures.py \
 	tests/e2e/src/sqlbuild/cli/commands/main/build/test_audit_failures.py \
 	tests/e2e/src/sqlbuild/cli/commands/main/build/test_build.py \
 	tests/e2e/src/sqlbuild/cli/commands/main/build/test_compile_json_behavior.py \
