+++ src/sqlbuild/python_nodes/models.py
@@ -160,6 +160,8 @@ class CheckDefinition:
     group: str | None = None
     description: str | None = None
     meta: dict[str, object] | None = None
+
+
 @dataclass(frozen=True, kw_only=True)
 class AuditCase:
     """One named generic-audit attachment generated by an audit factory."""
